bit data typed removed
xc8 is not compatible to datatype "bit"
the code was change in order to read bit zero and compare it to the status register of the LM639.

Also, some calls of the function "check_busy" was missing the empty paranthesis, this was fixed.
the "htc.h" and "pic18f46k22.h" files were replaced by "xc.h".

A couple of comments were added